Optimized the RNG
Optimized the RNG Process to have everything generated as a long, then converted back to the original format.
