+++ RNG/RNG.cs
@@ -11,25 +11,12 @@ namespace MyRNG
         /*
         ===============================================================================================
         PURPOSE:
-        Enumerate the different data types being generated.
-        ===============================================================================================
-        */
-        private enum DataTypes
-        {
-            Byte,
-            Integer,
-            Long
-        }
-
-        /*
-        ===============================================================================================
-        PURPOSE:
         Enumerate the different processes used in generating a random number.
         -----------------------------------------------------------------------------------------------
         NOTES:
         - The selection "END" is used to dynamically identify how many selections we have.
         - It is assumed that the return value, MinValue and MaxValue of a generated number will all be
-          the same.
+          the same type.
         ===============================================================================================
         */
         private enum RNGTypes
